ACL/Data Structures cleanup.
* Remove panics in favour of error handling.
* Rearrange modules.
* Fix up docs (slightly).
