+++ kernel-ewasm/cap9-std/examples/acl_admin.rs
@@ -61,7 +61,7 @@ pub trait ACLAdminInterface {
             // This relies on a mapping of groups -> procedures. Therefore we
             // need a map mechanism. Here we will just create the mechanism each
             // time at the same address.
-            let mut procecedure_map: cap9_std::StorageEnumerableMap<u8,cap9_std::SysCallProcedureKey> = cap9_std::StorageEnumerableMap::from(1);
+            let mut procecedure_map: cap9_std::StorageEnumerableMap<u8,cap9_std::SysCallProcedureKey> = cap9_std::StorageEnumerableMap::from(1).unwrap();
             procecedure_map.insert(group_id.as_u32() as u8, proc_key.into());
         }
 
@@ -69,7 +69,7 @@ pub trait ACLAdminInterface {
             // This relies on a mapping of groups -> procedures. Therefore we
             // need a map mechanism. Here we will just create the mechanism each
             // time at the same address.
-            let procecedure_map: cap9_std::StorageEnumerableMap<u8,cap9_std::SysCallProcedureKey> = cap9_std::StorageEnumerableMap::from(1);
+            let procecedure_map: cap9_std::StorageEnumerableMap<u8,cap9_std::SysCallProcedureKey> = cap9_std::StorageEnumerableMap::from(1).unwrap();
             match procecedure_map.get(group_id.as_u32() as u8) {
                 Some(x) => x.into(),
                 None => H256::zero(),
@@ -77,12 +77,12 @@ pub trait ACLAdminInterface {
         }
 
         fn set_account_group(&mut self, account: Address, group_id: U256) {
-            let mut procecedure_map: cap9_std::StorageEnumerableMap<Address, u8> = cap9_std::StorageEnumerableMap::from(0);
+            let mut procecedure_map: cap9_std::StorageEnumerableMap<Address, u8> = cap9_std::StorageEnumerableMap::from(0).unwrap();
             procecedure_map.insert(account, group_id.as_u32() as u8);
         }
 
         fn get_account_group(&mut self, account: Address) -> U256 {
-            let procecedure_map: cap9_std::StorageEnumerableMap<Address, u8> = cap9_std::StorageEnumerableMap::from(0);
+            let procecedure_map: cap9_std::StorageEnumerableMap<Address, u8> = cap9_std::StorageEnumerableMap::from(0).unwrap();
             match procecedure_map.get(account) {
                 Some(x) => x.into(),
                 None => U256::zero(),
@@ -139,7 +139,7 @@ pub trait ACLAdminInterface {
         fn proxy(&mut self, payload: Vec<u8>) {
             let sender = pwasm_ethereum::origin();
             let group_id = self.get_account_group(sender).as_u32() as u8;
-            let procecedure_map: cap9_std::StorageEnumerableMap<u8,cap9_std::SysCallProcedureKey> = cap9_std::StorageEnumerableMap::from(0);
+            let procecedure_map: cap9_std::StorageEnumerableMap<u8,cap9_std::SysCallProcedureKey> = cap9_std::StorageEnumerableMap::from(0).unwrap();
             let procedure_key = procecedure_map.get(group_id).unwrap();
             // Here the cap is hard coded. This procedure expects its first
             // procedure call capability to give it all the necessary
